backend/pyproject.toml
[project]
name = "chat-ui-backend"
version = "0.1.0"
description = "FastAPI backend for Chat UI"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openai-chatkit>=0.1.0",
    "openai-agents>=0.3.2",
]
